# supplementary_files/swagger documentation.yaml
swagger: '2.0'
basePath: /
paths:
...
  /api/chat_history:
    get:
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ChatHistory'
      summary: Retrieve the user's chat history for the current session
      operationId: get_chat_history
      parameters:
      - name: X-Fields
        in: header
        type: string
        format: mask
        description: An optional fields mask
      security:
      - BearerAuth: []
      tags:
      - chatbot
  /api/chatbot_stats:
    get:
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ChatbotStats'
      summary: Retrieve expanded, overall statistics about chatbot usage
      operationId: get_chatbot_stats
      parameters:
      - name: X-Fields
        in: header
        type: string
        format: mask
        description: An optional fields mask
      security:
      - BearerAuth: []
      tags:
      - chatbot
...
  /api/send_message:
    post:
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ChatResponse'
      summary: Send a message to the chatbot and get a reply
      operationId: send_message
      parameters:
      - name: payload
        required: true
        in: body
        schema:
          $ref: '#/definitions/ChatRequest'
      - name: X-Fields
        in: header
        type: string
        format: mask
        description: An optional fields mask
      security:
      - BearerAuth: []
      tags:
      - chatbot
...
info:
  title: GrowUp Richie Swagger API Documentation
  version: '1.0'
  description: API for a financial literacy app for children aged 8-14
produces:
- application/json
consumes:
- application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: Authorization
    description: Enter your JWT token with the `Bearer ` prefix, e.g., "Bearer <your_token>"
security:
- BearerAuth: []
...
definitions:
...
  ChatRequest:
    required:
    - message
    properties:
      message:
        type: string
        description: The user's message to the chatbot
    type: object
  ChatResponse:
    properties:
      reply:
        description: The bot's reply
        allOf:
        - $ref: '#/definitions/Message'
    type: object
  Message:
    properties:
      sender:
        type: string
        description: Who sent the message ('user' or 'bot')
      text:
        type: string
        description: The content of the message
      timestamp:
        type: string
        description: The time the message was sent
    type: object
  ChatHistory:
    properties:
      history:
        type: array
        description: The user's chat history
        items:
          $ref: '#/definitions/Message'
    type: object
  ChatbotStats:
    properties:
      total_messages:
        type: integer
        description: Total number of messages (user and bot).
      total_user_messages:
        type: integer
        description: Total number of messages sent by users.
      total_bot_responses:
        type: integer
        description: Total number of responses sent by the bot.
      unique_users_chatted:
        type: integer
        description: Total number of unique users who have used the chatbot.
      premium_users_chatted:
        type: integer
        description: Number of unique premium users who have used the chatbot.
      non_premium_users_chatted:
        type: integer
        description: Number of unique non-premium users who have used the chatbot.
      active_users_today:
        type: integer
        description: Number of unique users who used the chatbot today.
      active_users_last_7_days:
        type: integer
        description: Number of unique users who used the chatbot in the last 7 days.
      avg_messages_per_user:
        type: number
        description: Overall average number of messages sent per user.
      avg_messages_per_session:
        type: number
        description: Overall average number of messages per chat session.
    type: object
...